refactor(api): move csv export for profile collection into its bounded context
