+++ api/lib/infrastructure/repositories/campaign-participation-repository.js
@@ -54,26 +54,6 @@ const update = async function (campaignParticipation, domainTransaction = Domain
   await knexConn.from('campaign-participations').where({ id: campaignParticipation.id }).update(attributes);
 };
 
-const findProfilesCollectionResultDataByCampaignId = async function (campaignId) {
-  const results = await knex('campaign-participations')
-    .select([
-      'campaign-participations.*',
-      'view-active-organization-learners.studentNumber',
-      'view-active-organization-learners.division',
-      'view-active-organization-learners.group',
-      'view-active-organization-learners.firstName',
-      'view-active-organization-learners.lastName',
-    ])
-    .join(
-      'view-active-organization-learners',
-      'view-active-organization-learners.id',
-      'campaign-participations.organizationLearnerId',
-    )
-    .where({ campaignId, isImproved: false, 'campaign-participations.deletedAt': null });
-
-  return results.map(_rowToResult);
-};
-
 const findLatestOngoingByUserId = async function (userId) {
   const campaignParticipations = await knex('campaign-participations')
     .join('campaigns', 'campaigns.id', 'campaign-participations.campaignId')
@@ -214,7 +194,6 @@ export {
   getCodeOfLastParticipationToProfilesCollectionCampaignForUser,
   get,
   update,
-  findProfilesCollectionResultDataByCampaignId,
   findLatestOngoingByUserId,
   findOneByCampaignIdAndUserId,
   updateWithSnapshot,
@@ -224,24 +203,6 @@ export {
   getAllCampaignParticipationsInCampaignForASameLearner,
   remove,
 };
-
-function _rowToResult(row) {
-  return {
-    id: row.id,
-    createdAt: new Date(row.createdAt),
-    isShared: row.status === CampaignParticipationStatuses.SHARED,
-    sharedAt: row.sharedAt ? new Date(row.sharedAt) : null,
-    participantExternalId: row.participantExternalId,
-    userId: row.userId,
-    isCompleted: row.state === 'completed',
-    studentNumber: row.studentNumber,
-    participantFirstName: row.firstName,
-    participantLastName: row.lastName,
-    division: row.division,
-    pixScore: row.pixScore,
-    group: row.group,
-  };
-}
 
 function _getAttributes(campaignParticipation) {
   return {
