+++ api/src/prescription/campaign/application/campaign-detail-route.js
@@ -37,6 +37,25 @@ const register = async function (server) {
         ],
       },
     },
+    {
+      method: 'GET',
+      path: '/api/campaigns/{id}/csv-profiles-collection-results',
+      config: {
+        auth: false,
+        validate: {
+          params: Joi.object({
+            id: identifiersType.campaignId,
+          }),
+        },
+        handler: campaignDetailController.getCsvProfilesCollectionResults,
+        notes: [
+          "- **Cette route est restreinte via un token dédié passé en paramètre avec l'id de l'utilisateur.**\n" +
+            "- Récupération d'un CSV avec les résultats de la campagne de collecte de profils\n" +
+            '- L‘utilisateur doit avoir les droits d‘accès à l‘organisation liée à la campagne à créer',
+        ],
+        tags: ['api', 'campaign'],
+      },
+    },
   ]);
 };
