+++ api/tests/acceptance/application/campaigns/campaign-controller_test.js
@@ -277,95 +277,6 @@ describe('Acceptance | API | Campaign Controller', function () {
     });
   });
 
-  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
-    let accessToken;
-
-    function _createTokenWithAccessId({ userId, campaignId }) {
-      return jwt.sign(
-        {
-          access_id: userId,
-          campaign_id: campaignId,
-        },
-        settings.authentication.secret,
-        { expiresIn: settings.authentication.accessTokenLifespanMs },
-      );
-    }
-
-    beforeEach(async function () {
-      const userId = databaseBuilder.factory.buildUser().id;
-      organization = databaseBuilder.factory.buildOrganization();
-      campaign = databaseBuilder.factory.buildCampaign({
-        organizationId: organization.id,
-        type: 'PROFILES_COLLECTION',
-      });
-      accessToken = _createTokenWithAccessId({ userId, campaignId: campaign.id });
-
-      databaseBuilder.factory.buildMembership({
-        userId,
-        organizationId: organization.id,
-        organizationRole: Membership.roles.MEMBER,
-      });
-
-      await databaseBuilder.commit();
-
-      const learningContent = [
-        {
-          id: 'recArea1',
-          title_i18n: {
-            fr: 'area1_Title',
-          },
-          color: 'specialColor',
-          competences: [
-            {
-              id: 'recCompetence1',
-              name_i18n: { fr: 'Fabriquer un meuble' },
-              index: '1.1',
-              tubes: [
-                {
-                  id: 'recTube1',
-                  skills: [
-                    {
-                      id: 'recSkill1',
-                      nom: '@web2',
-                      challenges: [],
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        },
-      ];
-
-      const learningContentObjects = learningContentBuilder.fromAreas(learningContent);
-      mockLearningContent(learningContentObjects);
-    });
-
-    it('should return csv file with statusCode 200', async function () {
-      // given & when
-      const { statusCode, payload } = await server.inject({
-        method: 'GET',
-        url: `/api/campaigns/${campaign.id}/csv-profiles-collection-results?accessToken=${accessToken}`,
-      });
-
-      // then
-      expect(statusCode).to.equal(200, payload);
-    });
-
-    context('when accessing another campaign with the wrong access token', function () {
-      it('should return an error response with an HTTP status code 403', async function () {
-        // given & when
-        const { statusCode } = await server.inject({
-          method: 'GET',
-          url: `/api/campaigns/1234567890/csv-profiles-collection-results?accessToken=${accessToken}`,
-        });
-
-        // then
-        expect(statusCode).to.equal(403);
-      });
-    });
-  });
-
   describe('GET /api/campaigns/{id}/analyses', function () {
     let userId;
