+++ api/tests/prescription/campaign/acceptance/application/campaign-detail-controller_test.js
@@ -1,3 +1,5 @@
+import jwt from 'jsonwebtoken';
+
 import {
   databaseBuilder,
   expect,
@@ -6,10 +8,14 @@ import {
   mockLearningContent,
 } from '../../../../test-helper.js';
 
+import { config as settings } from '../../../../../lib/config.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
 import { createServer } from '../../../../../server.js';
 
 describe('Acceptance | API | Campaign Detail Controller', function () {
   let server;
+  let campaign;
+  let organization;
 
   beforeEach(async function () {
     server = await createServer();
@@ -67,6 +73,95 @@ describe('Acceptance | API | Campaign Detail Controller', function () {
 
       // then
       expect(response.statusCode).to.equal(200);
+    });
+  });
+
+  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
+    let accessToken;
+
+    function _createTokenWithAccessId({ userId, campaignId }) {
+      return jwt.sign(
+        {
+          access_id: userId,
+          campaign_id: campaignId,
+        },
+        settings.authentication.secret,
+        { expiresIn: settings.authentication.accessTokenLifespanMs },
+      );
+    }
+
+    beforeEach(async function () {
+      const userId = databaseBuilder.factory.buildUser().id;
+      organization = databaseBuilder.factory.buildOrganization();
+      campaign = databaseBuilder.factory.buildCampaign({
+        organizationId: organization.id,
+        type: 'PROFILES_COLLECTION',
+      });
+      accessToken = _createTokenWithAccessId({ userId, campaignId: campaign.id });
+
+      databaseBuilder.factory.buildMembership({
+        userId,
+        organizationId: organization.id,
+        organizationRole: Membership.roles.MEMBER,
+      });
+
+      await databaseBuilder.commit();
+
+      const learningContent = [
+        {
+          id: 'recArea1',
+          title_i18n: {
+            fr: 'area1_Title',
+          },
+          color: 'specialColor',
+          competences: [
+            {
+              id: 'recCompetence1',
+              name_i18n: { fr: 'Fabriquer un meuble' },
+              index: '1.1',
+              tubes: [
+                {
+                  id: 'recTube1',
+                  skills: [
+                    {
+                      id: 'recSkill1',
+                      nom: '@web2',
+                      challenges: [],
+                    },
+                  ],
+                },
+              ],
+            },
+          ],
+        },
+      ];
+
+      const learningContentObjects = learningContentBuilder.fromAreas(learningContent);
+      mockLearningContent(learningContentObjects);
+    });
+
+    it('should return csv file with statusCode 200', async function () {
+      // given & when
+      const { statusCode, payload } = await server.inject({
+        method: 'GET',
+        url: `/api/campaigns/${campaign.id}/csv-profiles-collection-results?accessToken=${accessToken}`,
+      });
+
+      // then
+      expect(statusCode).to.equal(200, payload);
+    });
+
+    context('when accessing another campaign with the wrong access token', function () {
+      it('should return an error response with an HTTP status code 403', async function () {
+        // given & when
+        const { statusCode } = await server.inject({
+          method: 'GET',
+          url: `/api/campaigns/1234567890/csv-profiles-collection-results?accessToken=${accessToken}`,
+        });
+
+        // then
+        expect(statusCode).to.equal(403);
+      });
     });
   });
 });
