+++ api/tests/prescription/campaign/unit/application/campaign-detail-route_test.js
@@ -47,4 +47,33 @@ describe('Unit | Application | Router | campaign-detail-router ', function () {
       expect(response.statusCode).to.equal(400);
     });
   });
+
+  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
+    it('should return 200', async function () {
+      // given
+      sinon
+        .stub(campaignDetailController, 'getCsvProfilesCollectionResults')
+        .callsFake((_, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns/1/csv-profiles-collection-results');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns/invalid/csv-profiles-collection-results');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
 });
