+++ api/tests/unit/application/campaign/index_test.js
@@ -65,35 +65,6 @@ describe('Unit | Application | Router | campaign-router ', function () {
     });
   });
 
-  describe('GET /api/campaigns/{id}/csv-profiles-collection-results', function () {
-    it('should return 200', async function () {
-      // given
-      sinon
-        .stub(campaignController, 'getCsvProfilesCollectionResults')
-        .callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/1/csv-profiles-collection-results');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 400 with an invalid campaign id', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/invalid/csv-profiles-collection-results');
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-
   describe('PATCH /api/admin/campaigns/{id}', function () {
     it('should return 204', async function () {
       // given
